requirements.txt
python-dotenv
qdrant-client
sentence-transformers
openai
requests
pdfplumber
tabulate
